The following were made to all special pvdbcr modules 1) Much cleaner implementation. 2) pvdbSpecialRegister.dbd is new
