+++ src/special/Makefile
@@ -2,10 +2,6 @@
 
 SRC_DIRS += $(PVDATABASE_SRC)/special
 
-LIBSRCS += pvdbcrTraceRecord.cpp
-LIBSRCS += pvdbcrRemoveRecord.cpp
-LIBSRCS += pvdbcrAddRecord.cpp
-LIBSRCS += pvdbcrProcessRecord.cpp
 
 DBD += pvdbcrTraceRecordRegister.dbd
 DBD += pvdbcrRemoveRecordRegister.dbd
@@ -13,6 +9,7 @@ DBD += pvdbcrAddRecordRegister.dbd
 DBD += pvdbcrProcessRecordRegister.dbd
 DBD += pvdbcrScalarRegister.dbd
 DBD += pvdbcrScalarArrayRegister.dbd
+DBD += pvdbSpecialRegister.dbd
 
 LIBSRCS += pvdbcrTraceRecordRegister.cpp
 LIBSRCS += pvdbcrRemoveRecordRegister.cpp
