+++ src/special/pvdbSpecialRegister.dbd
@@ -1,0 +1,6 @@
+include "pvdbcrAddRecordRegister.dbd"
+include "pvdbcrRemoveRecordRegister.dbd"
+include "pvdbcrProcessRecordRegister.dbd"
+include "pvdbcrTraceRecordRegister.dbd"
+include "pvdbcrScalarRegister.dbd"
+include "pvdbcrScalarArrayRegister.dbd"
